fix(details dialog): extract loading skeleton
